requirements.txt
tensorflow_gpu>=1.8.0
seaborn>=0.8.1
numpy>=1.14.3
gensim>=3.4.0
pandas>=0.22.0
nltk>=3.2.5
Keras>=2.1.6
keras_resnet>=0.1.0
joblib>=0.11
matplotlib>=2.2.2
pyenchant>=2.0.0
scikit_learn>=0.19.1
